+++ features/app/src/main/java/com/jlndev/facaseupedido/ui/uitls/model/ProductItem.kt
@@ -11,5 +11,5 @@ data class ProductItem(
     val description: String,
     val price: Double,
     val image: String,
-    var quantity: Int = 1
+    var quantity: Double = 1.0
 ): Parcelable, BaseDiffItemView()
